Popup menus have indexed entries
The plan is that you'll be able to select say the 12th entry by doing <C-!>
<C-@><C-N> or something like that.

Right now the index of the entry is stored in a variable
w:popupmenulastselectedindex, which will be useful for key mappings checking
pumvisible(), if keymappings are powerful enough. Perhaps we need to add some
functionality in the source to make this possible, hopefully not.

Still have to make it conditional based on some options.